backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nifty50-option-chain-backend"
version = "0.1.0"
description = "FastAPI backend for a Nifty 50 live option chain dashboard"
requires-python = ">=3.11"
dependencies = [
  "asyncpg>=0.30.0,<1.0.0",
  "email-validator>=1.1.0,<2.2",
  "fastapi>=0.115.0,<1.0.0",
  "fastapi-users[oauth,sqlalchemy]>=12.1.3,<13.0.0",
  "httpx>=0.28.0,<1.0.0",
  "orjson>=3.10.0,<4.0.0",
  "pydantic-settings>=2.7.0,<3.0.0",
  "python-multipart==0.0.7",
  "redis>=5.2.0,<6.0.0",
  "setuptools<81",
  "sqlalchemy>=2.0.36,<3.0.0",
  "uvicorn[standard]>=0.34.0,<1.0.0"
]

[tool.setuptools.packages.find]
include = ["app*"]
